SDIT-2873 Sequence doesn't need a Long (#1331)

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/nomisprisonerapi/jpa/OffenderExternalMovement.kt
@@ -31,7 +31,7 @@ data class OffenderExternalMovementId(
   val offenderBooking: OffenderBooking,
 
   @Column(name = "MOVEMENT_SEQ", nullable = false)
-  val sequence: Long,
+  val sequence: Int,
 ) : Serializable
 
 @EntityOpen
